pyproject.toml
[tool.poetry]
name = "tpay-payment"
version = "0.1.0"
description = ""
authors = ["BuildTools <unconfigured@null.spigotmc.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
aiogram = "2.25"
loguru = "^0.7.2"
httpx = "^0.27.2"
pydantic = "^2.10.2"
python-dotenv = "^1.0.1"
peewee-async = "0.8.1"
aiopg = "^1.4.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
